requirements.txt
absl-py==0.12.0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.5b1
Brotli==1.0.9
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.0.7
Django==3.2
django-plotly-dash==1.6.3
django-redis==4.12.1
docopt==0.6.2
dpd-components==0.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.9.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
hiredis==2.0.0
huggingface-hub==0.0.8
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
Js2Py==0.71
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
msgpack==1.0.2
mypy-extensions==0.4.3
numba==0.53.1
numpy==1.20.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
parso==0.8.2
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
prompt-toolkit==3.0.18
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyjsparser==2.7.1
pynndescent==0.5.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
service-identity==18.1.0
shap==0.39.0
six==1.15.0
slicer==0.0.7
soupsieve==2.2.1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
tqdm==4.60.0
traitlets==5.0.5
transformers==4.6.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
umap-learn==0.5.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.4.0
